+++ tests/testthat/test-easyCHAMP-particles.R
@@ -1,12 +1,12 @@
 test_that("Default processing works", {
-  expect_equal(evalPurency.particles(path="//HERE/COMES/YOUR/PATH/", 
+  expect_equal(easyCHAMP.particles(path="//HERE/COMES/YOUR/PATH/", 
                                       dataReturn = TRUE, 
                                      test = TRUE), 
                testdata.particles.default)
 })
 
 test_that("Processing with colourSep works", {
-  expect_equal(evalPurency.particles(path="//HERE/COMES/YOUR/PATH/", 
+  expect_equal(easyCHAMP.particles(path="//HERE/COMES/YOUR/PATH/", 
                                      colourSep = "transparent",
                                      dataReturn = TRUE, 
                                      test = TRUE), 
@@ -15,7 +15,7 @@ test_that("Processing with colourSep works", {
 })
 
 test_that("Processing without blank works", {
-  expect_equal(evalPurency.particles(path="//HERE/COMES/YOUR/PATH/", 
+  expect_equal(easyCHAMP.particles(path="//HERE/COMES/YOUR/PATH/", 
                                      dataReturn = TRUE, 
                                      noBlank = TRUE,
                                      test = TRUE),